test: move cdp only tests to a subfolder
And allow filtering to exclude cdp only tests.
